api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-pubsub==2.27.0
google-cloud-bigquery==3.27.0
pydantic==2.9.2
